storage/requirements.txt
Pillow==8.4.0
opencv-python==4.5.4.60
numpy==1.21.4
pandas==1.3.4
boto3==1.20.26
scipy==1.7.3
matplotlib==3.5.0
python-dotenv==0.19.2
tqdm==4.62.3